remove graph prep, as it breaks local PO assumptions
Each consensus subgraph (the POA block) is assumed to be partially
ordered in the sort of the graph given to consensus_graph.

Sorting the graph to "clean it up" was introducing local nonlinearity
into the sort. Removing this resolves a problem in consensus generation,
where many nodes were created for regions where one path was inverted
relative to the rest.