remove funky looping parts of the consensus graph
The consensus graph algorithm was introducing strage loops with many
embedded small variants. After resolving several problems, some cases of
these paths still existed.

These regions tended to have no link paths covering them! These could be
resolved with odgi prune and odgi unchop, but not when this was run
inline inside of smoothxg!

As with the memory fragmentation issue in smooth.cpp, this can be fixed
by copying the graph, and then doing unchop on the resulting graph.

The issue appears to be with the destruction of paths. Some steps remain
on nodes even if the path has been removed. This makes the coverage
filter (equivalent to odgi prune) not work properly. This commit is a
hot-fix, and a test case for odgi should be used to resolve this
problem.

But, it is a fix for the complex consensus graph regions.